chore: add weighted tests to ce and fix tests
